[rom] explain why it's hard to push the maths pack into an expansion ROM
TL;DR - the floating point routines don't use RAM LAM, so they will
read from a ROM if the upper expansion ROM is enabled.  So, if code
in an expansion ROM calls these routines, it'll flip back the
expansion ROM and .. well, bad things happen.  We'd need a version
of RAM LAM that used the PREVIOUS ROM config to read/copy bytes from.